moving plugins into their own package
